dom0: remove xenstore initialization from application main()
Previously, Xenstore root nodes initialization was placed inside the
application main function. It was a big hack, because now zephyr-xenlib
where actual xenstore-srv is placed, is now separate project and we
shouldn't initialize it in such way. Now xenstore-srv is fixed and root
initialization is performed via SYS_INIT on APPLICATION stage, so we
removing it from zephyr-dom0-xt.

Signed-off-by: Dmytro Firsov <dmytro_firsov@epam.com>
Acked-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>
Acked-by: Leonid Komarianskyi <leonid_komarianskyi@epam.com>
Acked-by: Mykola Kvach <mykola_kvach@epam.com>