prj.conf: adjust shell stack size for application
Previously this size was set by board configs, which is not correct
in Zephyr config conception. Now, with new board config it is set to
default size of 1KB, which is not enough for our setup. Set it to 4KB,
value was picked by tests, it should be enough for our purposes.

Signed-off-by: Dmytro Firsov <dmytro_firsov@epam.com>